Fix default sysvar cache loading and add slot bank slot
